+++ dwds/CHANGELOG.md
@@ -1,4 +1,8 @@
-## 18.0.1-dev
+## 18.0.2-dev
+
+- Support new DDC temp names for patterns. - [#2042](https://github.com/dart-lang/webdev/pull/2042)
+
+## 18.0.1
 
 - Fix failure to map JS exceptions to dart. - [#2004](https://github.com/dart-lang/webdev/pull/2004)
 - Fix for listening to custom streams. - [#2011](https://github.com/dart-lang/webdev/pull/2011)