+++ webdev/CHANGELOG.md
@@ -1,4 +1,8 @@
-## 3.0.2-dev
+## 3.0.3-dev
+
+## 3.0.2
+
+-  Update `build_daemon` constraint to `^4.0.0`.
 
 ## 3.0.1